requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.4
PyMySQL==1.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
